% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictonary.R
\name{add_sysdic}
\alias{add_sysdic}
\title{Add user-defined dictionary files to system dictionary.}
\usage{
add_sysdic()
}
\description{
사용자가 정의한 사용자 정의 사전 파일을 시스템 사전에 추가
}
\details{
사용자 사전정의 디렉토리에 있는 모든 사용자정의 사전 파일을 시스템 사전에 추가한다.
}
\examples{
\dontrun{
add_sysdic()
}
}
