% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictonary.R
\name{add_userdic}
\alias{add_userdic}
\title{Add user-defined dictionary files to user dictionary.}
\usage{
add_userdic()
}
\description{
사용자가 정의한 사용자 정의 사전 파일을 사용자 사전에 추가
}
\details{
사용자 사전정의 디렉토리에 있는 모든 사용자정의 사전 파일을 사용자 사전에 추가한다.
}
\examples{
\dontrun{
add_userdic()
}
}
