% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictonary.R
\name{append_userdic_nnp}
\alias{append_userdic_nnp}
\title{Write to the user-defined person dictionary file.}
\usage{
append_userdic_nnp(x, type = c("person", "place", "nnp"), userdic_path = NULL)
}
\arguments{
\item{x}{character. 고유명사 사전에 등록할 이름들.}

\item{type}{character. 인명사전과 지명사전, 고유명사 사전에서 등록할 사용자 정의 고유명사 사전 선택.}

\item{userdic_path}{character. 사용자 정의 고유명사 사전 파일이 존재하는 경로.
지정하지 않으면 사전이 설치된 기본 경로에서 파일을 읽어온다.}
}
\description{
사용자 사전 중 고유명사 사전에 등록하기 위해 인명/지명을 인명/지명/고유명사 사전 파일에 추가
}
\details{
사용자 사전정의 디렉토리의 person.csv/place.csv/nnp.csv 파일에 등록할 인명/지명/고유명사를 추가한다.
}
\examples{
\dontrun{
# 인명 사전
get_userdic_nnp()
append_userdic_nnp(c("변학도"))

# 지명 사전
get_userdic_nnp("place")
append_userdic_nnp(c("영귀미면"), "place")
get_userdic_nnp("place")

# 기타 고유명사 사전  
get_userdic_nnp("nnp")
append_userdic_nnp(c("릴리움"), "nnp")
get_userdic_nnp("nnp")
}
}
