% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictonary.R
\name{append_userdic_person}
\alias{append_userdic_person}
\title{Write to the user-defined person dictionary file.}
\usage{
append_userdic_person(x, userdic_path = NULL)
}
\arguments{
\item{x}{character. 인명 사전에 등록할 이름들.}

\item{userdic_path}{character. 사용자 정의 인명사전 파일이 존재하는 경로.
지정하지 않으면 사전이 설치된 기본 경로에서 파일을 읽어온다.}
}
\description{
사용자 사전 중 인명사전에 등록하기 위해 인명을 인명사전 파일에 추가
}
\details{
사용자 사전정의 디렉토리의 person.csv 파일에 등록할 인명을 추가한다.
}
\examples{
\dontrun{
get_userdic_person()
append_userdic_person(c("변학도"))
get_userdic_person()
}
}
