% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collocate.R
\name{coll_scores}
\alias{coll_scores}
\title{Calculate t-score and mutual information score}
\usage{
coll_scores(x, node, span = 3)
}
\arguments{
\item{x}{data.frame. "collocate()"를 수행한 공동발생(co-occurrences) 분석결과}

\item{node}{character. 공동발생 분석 단어(term)}

\item{span}{integer. 공동발생 window 단위. 기본값은 3.}
}
\value{
data.frame. 공동발생 정보와 T-score, MI-score를 담은 data.frame
}
\description{
공동발생 분석을 위한 공동발생 단어에 대한 t-score, MI(mutual information)-score 계산
}
\examples{
\donttest{
docs <- president_speech$doc[1]

# default arguments
collocate(docs, "우정", type = "morpheme")

# change span argument
tab_colloc <- collocate(docs, "우정", type = "morpheme") 
coll_scores(tab_colloc, "우정", span = 2)

# change span argument
tab_colloc <- collocate(docs, "우정", type = "morpheme") 
coll_scores(tab_colloc, "국민", span = 2)
}
}
