% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{collapse_noun}
\alias{collapse_noun}
\title{Extract Collapsed Noun}
\usage{
collapse_noun(
  doc,
  user_dic = NULL,
  type = c("noun", "noun2"),
  chunk = round(length(if (tibble::is_tibble(doc)) dplyr::pull(doc) else doc)/mc.cores),
  mc.cores = parallel::detectCores()
)
}
\arguments{
\item{doc}{character. 명사로만 구성될 문서를 만들 대상 텍스트 데이터}

\item{user_dic}{mecab-ko 형태소 분석기의 사용자 정의 사전 파일.
기본값은 NULL로 사용자 사전파일을 지정하지 않음.
시스템 사전인 "/usr/local/lib/mecab/dic/mecab-ko-dic"(Linux, Mac)를 보완하여 사용됨.
사용자 사전 파일은 mecab-dict-index 명령어로 생성되며, 확장자가 "dic"임.}

\item{type}{character. 토큰화할 명사의 유형을 지정. "noun", "noun2"중에서 선택. 
기본값은 "noun"로 일반명사를 의미하며, "noun2"는 모든 명사를 의미함.}

\item{chunk}{integer. 병렬 작업 수행 시 처리 단위인 chunk}

\item{mc.cores}{integer. 병렬 작업 수행 시 사용할 코어의 개수}
}
\value{
character. 명사로만 구성된 텍스트
}
\description{
텍스트 문서에서 명사들을 토큰화한 후, 토큰화된 명사들을 공백으로
묶어서 텍스트 문서를 만듦
조회한다.
}
\details{
MS-Windows에서는 병렬처리를 지원하지 않음
}
\examples{
\donttest{
collapse_noun(president_speech$doc[1:7])

# Collaboration with tidytext
library(dplyr)

nho_noun <- president_speech \%>\%
  filter(president \%in\% "노무현") \%>\%
  filter(stringr::str_detect(category, "^외교")) \%>\%
  mutate(doc_noun = collapse_noun(doc)) \%>\%
    tidytext::unnest_ngrams(
      noun_bigram,
      doc_noun,
      n = 2
   )
nho_noun
 
nho_noun$noun_bigram[1:5]
}
}
