% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collocate.R
\name{collocate}
\alias{collocate}
\title{Calculate table for co-occurrence analysis}
\usage{
collocate(
  x,
  node,
  span = 3,
  type = c("noun", "noun2", "verb", "adj", "morpheme")
)
}
\arguments{
\item{x}{character. 공동발생(co-occurrences) 분석에 사용할 document.}

\item{node}{character. 공동발생 분석한 단어(term)}

\item{span}{integer. 공동발생 window 단위. 기본값은 3.}

\item{type}{character. 공동발생에 사용할 단어를 생성하는 방법으로서의 형태소 분석의 결과 유형. 
모든 품사, 명사, 동사 및 형용사와 같은 토큰화 결과 유형을 지정.
 "morpheme", "noun", "noun2", "verb", "adj"중에서 선택. 기본값은 "noun"로
 일반명사만 추출함.}
}
\value{
data.frame 공동발생 정보를 담은 data.frame
}
\description{
공동발생 분석을 위한 공동발생 단어 추출 및 해당 단어의 공동발생 빈도 및 문서에서의 발생 빈도 정보 생성
}
\examples{
\donttest{
docs <- president_speech$doc[1]

# default arguments
collocate(docs, "우정")

# change span argument
collocate(docs, "우정", span = 4)

# change type argument
collocate(docs, "우정", type = "morpheme")
}
}
