% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictonary.R
\name{create_userdic}
\alias{create_userdic}
\title{create user dictionary with user-defined dictionary files.}
\usage{
create_userdic(
  userdic_path = "./user_dic",
  dic_file = "user-dic.dic",
  installd = NULL,
  dic_path = NULL
)
}
\description{
사용자가 정의한 사용자 정의 사전 파일을 사용자 사전으로 생성
}
\details{
사용자 사전정의 디렉토리에 있는 모든 사용자정의 사전 파일을 엮어 사용자 사전에 추가한다.
}
\examples{
\dontrun{
create_userdic()
}
}
