% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictonary.R
\name{edit_termcost}
\alias{edit_termcost}
\title{Modify the word cost of a word in a dictionary definition file.}
\usage{
edit_termcost(userdic_path = "./user_dic", dic_file = "user-dic.dic")
}
\arguments{
\item{userdic_path}{character. 사용자 정의 사전 파일이 존재하는 경로.
지정하지 않으면 "./user_dic"이라는 이름의 경로를 사용함.}

\item{dic_file}{character. 생성할 사용자 사전 파일 이름.
지정하지 않으면 "user-dic.dic"이라는 이름으로 생성함.}
}
\description{
사용자 정의 사전 파일에서 낱말의 낱말비용을 수정
}
\details{
사용자 사전정의 디렉토리에 있는 병합 사용자 사전 정의 파일의 낱말비용을 수정함.
대상 파일은 "사용자 정의 사전 파일 경로/indexed/merged.csv" 파일임. 
그리고 낱말비용이 수정되면 수정된 낱말의 정보는 "사용자 정의 사전 파일 경로/indexed/changed.csv" 
파일에 수정 이력이 저장됨. 이 정보는 사용자 사전 정의 파일을 재정의할 때 
수정된 낱말비용 정보가 없어지는 매커니즘을 보완하기 위한 이력 파일임.
낱말비용이 수정되었다는 것 만으로 그 정보가 사전에 반영되는 것은 아님.
}
\examples{
\dontrun{
edit_termcost()
}
}
