% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{explore_docs}
\alias{explore_docs}
\title{Text Data Explorer}
\usage{
explore_docs()
}
\value{
없음
}
\description{
정규표현식 기반의 텍스트 데이터  탐색 작업을 위한 Shiny 앱 호출
}
\examples{
\dontrun{
 library(bitNLP)

 ## 텍스트 데이터 탐색기(Shiny Web Application) 호출
 explore_docs()
}
}
\seealso{
\code{\link{morpho_mecab}}
}
\author{
유충현
Maintainer: 유충현 <choonghyun.ryu@gmail.com>
}
