% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{get_plan_cost}
\alias{get_plan_cost}
\title{Search for tokenizer plans based on word cost}
\usage{
get_plan_cost(x)
}
\arguments{
\item{x}{character. 플랜을 조회한 단어나 문장.}
}
\value{
tbl_df. 플랜을 담은 tibble 객체.
}
\description{
낱말비용 기반의 토크나이저 플랜을 조회한다.
}
\details{
mecab-ko가 주어진 단어나 문장을 토크화(tokenization)하는 플랜을 조회한다. 
이 기능을 통해서 사전에서의 단어 비용 조정과 신규 사용자 단어의 추가를 의사결정 할 수 있다.
우선 순위에 따른 10개의 플랜을 조회하며, 실제로 형태소분석기는 우선 순위가 1인 것으로 토큰화한다.
현재 이 함수는 Windows 운영체제를 지원하지 않는다.
플랜 정보에서 변수는 다음과 같다.:
\itemize{
\item "우선순위" : 토큰화 우선 순위.
\item "표층형" : 토큰화되는 토큰
\item "품사태그" : 토큰의 품사.
\item "의미부류" : 인명, 혹은 지명과 같은 의미.
\item "좌문맥ID" : 좌문맥 ID.
\item "우문맥ID" : 우문맥 ID.
\item "낱말비용" : 가중치. 값은 낮을수록 가중치가 올라간다.
\item "연접비용" : 좌측에 공백 문자를 포함하는 품사의 연접 비용.
\item "누적비용" : 누적 낱말비용
}
}
\examples{
\dontrun{
get_plan_cost("가면무도회")
}
}
