% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{get_spacing}
\alias{get_spacing}
\title{Korean automatic spacing}
\usage{
get_spacing(x, user_dic = NULL)
}
\arguments{
\item{x}{character. 띄어쓰기 보정에 사용할 document.}

\item{user_dic}{mecab-ko 형태소 분석기의 사용자 정의 사전 파일.
기본값은 NULL로 사용자 사전파일을 지정하지 않음.}
}
\value{
띄어쓰기 보정된 character 벡터.
}
\description{
한글 문장을 띄어쓰기 규칙에 맞게 자동으로 띄어쓰기 보정.
}
\examples{
\donttest{
# 한글 자동 띄어쓰기
get_spacing("최근음성인식정확도가높아짐에따라많은음성데이터가Text로변환되고분석되기시작했는데,이를위해잘동작하는띄어쓰기엔진은거의필수적인게되어버렸다")

str <- "글쓰기에서맞춤법과띄어쓰기를올바르게하는것은좋은글이될수있는요건중하나이다.하지만요즘학생들은부족한어문규정지식으로인해맞춤법과띄어쓰기에서많은오류를범하기도한다.본연구는그중띄어쓰기가글을인식하는데중요한역할을하는것으로판단하여,대학생들이띄어쓰기에대해서어느정도정확하게인식하고있는지,실제오류실태는어떠한지에대해살펴서그오류를개선할수있는교육방안을마련할필요가있다고판단하였다."
get_spacing(str)
}
}
