% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictonary.R
\name{get_userdic_noun}
\alias{get_userdic_noun}
\title{Query the user-defined person dictionary file.}
\usage{
get_userdic_noun(
  noun_type = c("person", "place", "nnp", "nng"),
  userdic_path = NULL
)
}
\arguments{
\item{noun_type}{character. 인명사전, 지명사전, 고유명사 사전, 일반명사 사전에서 
조회할 사용자 정의 명사 사전 선택. 기본값은 "person"로 인명사전을 지정함.}

\item{userdic_path}{character. 사용자 정의 명사 사전 파일이 존재하는 경로.
지정하지 않으면 사전이 설치된 기본 경로에서 파일을 읽어온다.}
}
\value{
spec_tbl_df. 명사 사전 정의를 담은 tibble 객체.
tibble 객체에서 변수는 다음과 같다.:
\itemize{
\item "표층형" : 낱말명.
\item "미지정1" : 사용하지 않는 컬럼.
\item "미지정2" : 사용하지 않는 컬럼.
\item "미지정3" : 사용하지 않는 컬럼.
\item "품사태그" : 인명의 품사. NNP를 사용함.
\item "의미부류" : 인명, 혹은 지명과 같은 의미 부류.
\item "종성유무" : 낱말의 마지막 음절의 종성 여부. T, F 입력.
\item "읽기" : 읽어서 소리나는 말.
\item "타입" : inflected, compound, Preanalysis, *.
\item "첫번째 품사" : 기분석으로 나눠지는 토큰에 대한 각 품사 입력.
\item "마지막 품사" : 기분석으로 나눠지는 토큰에 대한 각 품사 입력.
\item "표현" : 낱말이 토큰들로 나눠질 경우의 원형을 +로 묶어 입력
\item "인덱스표현" : 사용하지 않는 컬럼, *로 표현
}
}
\description{
사용자 사전 중에서 명사 사전의 내용을 조회한다.
}
\details{
사용자 사전정의 디렉토리의 사전파일 읽어, 정의된 내용을 tibble 객체로 반환한다. 
이 기능을 통해서 사용자 명사 사전의 등록(정의) 여부를 파악할 수 있다.
다음과 같은 명사 사용자 정의 사전 파일을 참조한다.
\itemize{
\item 인명사전 : person.csv
\item 지명사전 : place.csv
\item 고유명사사전 : nnp.csv
\item 일반명사사전 : nng.csv
}
인명, 지명, 고유명사, 일반명사 사전의 경우에는 타입, 첫번째 품사, 마지막 품사,
인텍스표현의 정보는 의미가 없어 모두 *로 표현함.
}
\examples{
\dontrun{
get_userdic_noun("person")
}
}
\references{
{
mecab-ko-dic 품사 태그 설명. 
<https://docs.google.com/spreadsheets/d/1-9blXKjtjeKZqsf4NzHeYJCrr49-nXeRF6D80udfcwY/edit#gid=1718487366>
}
}
