% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movie_ratings.R
\docType{data}
\name{movie_ratings_train}
\alias{movie_ratings_train}
\alias{movie_ratings_test}
\title{Naver sentiment movie corpus v1.0}
\format{
3개의 변수와 150,000개(train), 50,000(test)의 관측치로 구성된 티블(tibble) 객체.:
\describe{
  \item{id}{character. 리뷰 아이디}
  \item{document}{character. 영화 리뷰}
  \item{label}{integer. 긍부정의 정보. 부정(0), 긍정(1)}
}
}
\source{
"Naver sentiment movie corpus v1.0" in github <https://github.com/e9t/nsmc>
}
\usage{
data(movie_ratings_train)

data(movie_ratings_test)
}
\description{
네이버 영화 리뷰에서 스크랩한 데이터이며, 모두 140자 미만의 길이고, 
0(Negative)과 1(Positive)로 라벨링 되어있음
}
\examples{
\dontrun{
data(movie_ratings_train)
data(movie_ratings_test)

head(movie_ratings_train)
}
}
\keyword{datasets}
