% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_dic.R
\docType{data}
\name{sentiment_dic}
\alias{sentiment_dic}
\title{KNU Korean Sentiment Dictionary}
\format{
4개의 변수와 14,841개의 관측치로 구성된 데이터 프레임.:
\describe{
  \item{word}{character. 사전 단어}
  \item{word_root}{character. 어근}
  \item{polarity}{integer. 긍부정의 정보. 매우 부정(-2), 부정(-1), 중립(0), 긍정(1), 매우 긍정(2)}
  \item{n_gram}{integer. n-Gram 수}
}
}
\source{
"KNU 한국어 감성사전" in github <https://github.com/park1200656/KnuSentiLex>
}
\usage{
data(sentiment_dic)
}
\description{
2018년도 군산대학교 소프트웨어융합공학과 Data Intelligence Lab에서 개발한 
한국어 감성사전으로 총 14,841개의 1-gram, 2-gram, ..., 8-gram, 관용구, 문형, 축약어, 
이모티콘 등에 대한 긍정, 중립, 부정 판별 및 정도(degree)값 계산
}
\details{
표준국어대사전을 구성하는 각 단어의 뜻풀이를 분석하여 긍부정어를 추출하였으며, 표준국어대사전을 구성하는 형용사, 부사, 동사, 명사의 모든 뜻풀이에 대한 긍정, 중립, 부정으로 분류하기 위해 Bi-LSTM 딥 러닝 모델 사용
}
\examples{
\dontrun{
data(sentiment_dic)

head(sentiment_dic)
}
}
\keyword{datasets}
