% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{unnest_noun_ngrams}
\alias{unnest_noun_ngrams}
\title{Wrapper around unnest_tokens for n-grams of noun}
\usage{
unnest_noun_ngrams(
  tbl,
  output,
  input,
  n = 2L,
  n_min = n,
  ngram_delim = " ",
  drop = TRUE,
  collapse = NULL,
  ...
)
}
\arguments{
\item{tbl}{A data frame.}

\item{output}{character or symbol. 출력열로 새로 만들 변수 이름}

\item{input}{character or symbol. 입력으로 사용할 변수 이름}

\item{n}{integer. n-gram의 단어 수입니다. 1 이상의 정수. 기본값은 2.}

\item{n_min}{integer. 이것은 1보다 크거나 같고 n보다 작거나 같은 정수여야 함}

\item{ngram_delim}{character. 생성된 n-gram에서 단어 사이의 구분 기호}

\item{drop}{logical. 원래 입력 열을 삭제해야 하는지 여부. 기본값은 TRUE이며 
원래 입력 열과 새 출력 열의 이름이 같은 경우 무시됨.}

\item{collapse}{A character vector. 결과에서 개별 n-gram들을 그룹핑할 변수 이름. 
기본값은 NULL로 개별 n-gram들을 묶지 않음.}

\item{...}{토크나이저(tokenize_noun_ngrams)에 전달되는 추가 인수}

\item{stopwords}{character. n-그램에서 제외할 불용어의 문자형 벡터}
}
\value{
토큰화된 character 벡터를 성분으로 갖는 list. simplify값이 TRUE이고 
x가 단일값일 때에는 character 벡터
}
\description{
명사를 추출하여 n-gram으로 토큰화합니다.
}
\examples{
\donttest{
library(dplyr)

president_speech \%>\%
  select(title, doc) \%>\% 
  filter(row_number() <= 2) \%>\%
  unnest_noun_ngrams(
    noun_bigram,
    doc,
    n = 2,
    ngram_delim = ":",
    type = "noun2"
  )
  
president_speech \%>\%
  select(title, doc) \%>\% 
  filter(row_number() <= 2) \%>\%
  unnest_noun_ngrams(
    noun_bigram,
    doc,
    n = 2,
    ngram_delim = ":",
    drop = FALSE
  )   
 
# grouping using group_by() function
president_speech \%>\%
  filter(row_number() <= 4) \%>\%
  mutate(speech_year = substr(date, 1, 4)) \%>\% 
  select(speech_year, title, doc) \%>\% 
  group_by(speech_year) \%>\%
  unnest_noun_ngrams(
    noun_bigram,
    doc,
    n = 2,
    ngram_delim = ":"
  )
  
# grouping using collapse argument
president_speech \%>\%
  filter(row_number() <= 4) \%>\%
  mutate(speech_year = substr(date, 1, 4)) \%>\% 
  select(speech_year, title, doc) \%>\% 
  unnest_noun_ngrams(
    noun_bigram,
    doc,
    n = 2,
    ngram_delim = ":",
    collapse = "speech_year"
  )
}

}
