% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictonary.R
\name{create_userdic}
\alias{create_userdic}
\title{create user dictionary with user-defined dictionary files.}
\usage{
create_userdic(userdic_path = "./user_dic", dic_file = "user-dic.dic")
}
\arguments{
\item{userdic_path}{character. 사용자 정의 명사 사전 파일이 존재하는 경로.
지정하지 않으면 ./user_dic라는 이름의 경로를 사용함.}

\item{dic_file}{character. 생성할 사용자 사전 파일 이름.
지정하지 않으면 user-dic.dic라는 이름으로 생성함.}
}
\description{
사용자가 정의한 사용자 정의 사전 파일을 사용자 사전으로 생성
}
\details{
사용자 사전정의 디렉토리에 있는 모든 사용자정의 사전 파일을 엮어 사용자 사전에 추가한다.
}
\examples{
\dontrun{
create_userdic()
}
}
